YT_Audio_Downloader/requirements.txt
yt-dlp>=2023.12.30
customtkinter>=5.2.1
Pillow>=10.1.0
ffmpeg-python>=0.2.0